Introduce JsonSerializerOptionsHelper.
