+++ framework/src/Volo.Abp.Json/Volo/Abp/Json/SystemTextJson/JsonConverters/AbpExtraPropertyDictionaryJsonConverter.cs
@@ -13,8 +13,9 @@ namespace Volo.Abp.Json.SystemTextJson.JsonConverters
     {
         public override T Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
         {
-            var newOptions = new JsonSerializerOptions(options);
-            newOptions.Converters.RemoveAll(x => x == this || x.GetType() == typeof(AbpExtraPropertyDictionaryJsonConverterFactory));
+            var newOptions = JsonSerializerOptionsHelper.Create(options, x =>
+                x == this ||
+                x.GetType() == typeof(AbpExtraPropertyDictionaryJsonConverterFactory));
 
             var rootElement = JsonDocument.ParseValue(ref reader).RootElement;
             var extensibleObject = JsonSerializer.Deserialize<T>(rootElement.GetRawText(), newOptions);
@@ -31,8 +32,10 @@ namespace Volo.Abp.Json.SystemTextJson.JsonConverters
 
         public override void Write(Utf8JsonWriter writer, T value, JsonSerializerOptions options)
         {
-            var newOptions = new JsonSerializerOptions(options);
-            newOptions.Converters.RemoveAll(x => x == this || x.GetType() == typeof(AbpExtraPropertyDictionaryJsonConverterFactory));
+            var newOptions = JsonSerializerOptionsHelper.Create(options, x =>
+                x == this ||
+                x.GetType() == typeof(AbpExtraPropertyDictionaryJsonConverterFactory));
+
             JsonSerializer.Serialize(writer, value, newOptions);
         }
     }
