+++ framework/src/Volo.Abp.Json/Volo/Abp/Json/SystemTextJson/JsonConverters/AbpStringToEnumConverter.cs
@@ -28,16 +28,20 @@ namespace Volo.Abp.Json.SystemTextJson.JsonConverters
 
         public override T Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
         {
-            var newOptions = new JsonSerializerOptions(options);
-            newOptions.Converters.RemoveAll(x => x == this || x.GetType() == typeof(AbpStringToEnumFactory));
+            var newOptions = JsonSerializerOptionsHelper.Create(options, x =>
+                x == this ||
+                x.GetType() == typeof(AbpStringToEnumFactory));
+
             newOptions.Converters.Add(_innerJsonStringEnumConverter.CreateConverter(typeToConvert, newOptions));
             return JsonSerializer.Deserialize<T>(ref reader, newOptions);
         }
 
         public override void Write(Utf8JsonWriter writer, T value, JsonSerializerOptions options)
         {
-            var newOptions = new JsonSerializerOptions(options);
-            newOptions.Converters.RemoveAll(x => x == this || x.GetType() == typeof(AbpStringToEnumFactory));
+            var newOptions = JsonSerializerOptionsHelper.Create(options, x =>
+                x == this ||
+                x.GetType() == typeof(AbpStringToEnumFactory));
+
             JsonSerializer.Serialize(writer, value, newOptions);
         }
     }
