+++ framework/src/Volo.Abp.Json/Volo/Abp/Json/SystemTextJson/JsonConverters/ObjectToInferredTypesConverter.cs
@@ -51,8 +51,7 @@ namespace Volo.Abp.Json.SystemTextJson.JsonConverters
 
         public override void Write(Utf8JsonWriter writer, object objectToWrite, JsonSerializerOptions options)
         {
-            var newOptions = new JsonSerializerOptions(options);
-            newOptions.Converters.Remove(this);
+            var newOptions = JsonSerializerOptionsHelper.Create(options, this);
             JsonSerializer.Serialize(writer, objectToWrite, newOptions);
         }
     }
